Rename xml_utils to something more sensible
